chore: abstracts template functionality from OpenShiftAssistant.
